requirements.txt
alembic==1.11.3
amqp==5.1.1
apispec==6.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==3.3.1
Brotli==1.0.9
cachelib==0.4.1
cattrs==23.1.2
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
constantly==15.1.0
convertdate==2.4.0
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.3
daphne==4.0.0
Deprecated==1.2.14
deprecation==2.1.0
Django==4.2.4
django-extensions==3.2.3
django-filter==23.2
django-rest-framework==0.1.0
django-rest-framework-generate==0.4
django-schema-graph==3.1.0
djangorestframework==3.14.0
dnspython==2.4.2
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Flask==2.1.3
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-Compress==1.13
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.4.0
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-WTF==1.0.1
func-timeout==4.3.5
geographiclib==2.0
geopy==2.3.0
graphlib-backport==1.0.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
hashids==1.3.1
hijri-converter==2.3.1
holidays==0.17.2
humanfriendly==10.0
humanize==4.8.0
hyperlink==21.0.0
idna==3.4
importlib-resources==6.0.1
incremental==22.10.0
install==1.3.5
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
korean-lunar-calendar==0.3.1
limits==3.5.0
loguru==0.7.0
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.26.1
mdurl==0.1.2
mergedeep==1.3.4
msgpack==1.0.5
mysqlclient==2.2.0
numpy==1.25.2
openpyxl==3.1.2
ordered-set==4.1.0
outcome==1.2.0
packaging==23.1
pandas==1.5.3
paramiko==3.3.1
parsedatetime==2.6
pgsanity==0.2.9
polyline==2.0.0
postman2py==0.0.2
postpy2==0.0.8
prettyjson==1.0.3
prison==0.2.1
prompt-toolkit==3.0.39
psutil==5.9.5
pyarrow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
Pygments==2.16.1
PyJWT==2.8.0
pylibmagic==0.4.0
PyMeeus==0.5.12
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-geohash==0.8.5
python-magic-bin==0.4.14
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
selenium==4.11.2
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.4
sshtunnel==0.4.0
tabulate==0.8.10
trio==0.22.2
trio-websocket==0.10.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.7
whitenoise==6.5.0
wrapt==1.15.0
wsproto==1.2.0
WTForms==2.3.3
WTForms-JSON==0.3.5
XlsxWriter==3.0.9
zope.interface==6.0
